Update README.md
removed : - [In context translation with Respresso](https://blog.respresso.io/index.php/2020/06/22/in-context-translation-with-respresso/) - Real-time preview how the translations will look like in your mobile app or web. No need to wait for the next deployment.
This was in Reading and looks like it should either be a tool or nothing at all.

Otherwise, just reordered tools & articles into subcategories.